extract meta to types util
